nothing to see here

+++ GameService/src/main/java/nl/fontys/atosgame/gameservice/cardseeder/CardSeeder.java
@@ -34,6 +34,26 @@ public class CardSeeder {
     //When receiving a new card-set the old id's will always be thrown out, and it will insert new ones.
     public void handleCardSet(List<CardSet> currentCards) throws IOException
     {
+        List<String> idsOld = new ArrayList<>();
+        List<String> idsNew = new ArrayList<>();
+        for(CardSet set : oldCards)
+        {
+            idsOld.add(set.getId().toString());
+        }
+
+        for(CardSet set : currentCards)
+        {
+            idsNew.add(set.getId().toString());
+        }
+
+        Collections.sort(idsOld);
+        Collections.sort(idsNew);
+
+        if(idsOld.hashCode() == idsNew.hashCode())
+        {
+            return;
+        }
+
         for(CardSet set : oldCards)
         {
             cardSetService.deleteCardSet(set.getId());
